refactor: 랜덤 정수 생성기와 자동차 랜덤 정수 적용기 분리
