+++ src/main/java/racingcar/domain/ValidInputFromUserGetter.java
@@ -2,8 +2,8 @@ package racingcar.domain;
 
 import java.util.List;
 import java.util.Scanner;
-import racingcar.input.carname.ValidCarNamesInputFromUserGetter;
-import racingcar.input.racingtrytime.ValidRacingTryTimeFromUserGetter;
+import racingcar.domain.input.carname.ValidCarNamesInputFromUserGetter;
+import racingcar.domain.input.racingtrytime.ValidRacingTryTimeFromUserGetter;
 
 public class ValidInputFromUserGetter {
     private final ValidCarNamesInputFromUserGetter validCarNamesInputFromUserGetter;
