+++ src/main/java/racingcar/domain/input/racingtrytime/printer/RacingTryTimeInputFromUserPrinter.java
@@ -1,0 +1,7 @@
+package racingcar.domain.input.racingtrytime.printer;
+
+public class RacingTryTimeInputFromUserPrinter {
+    public void printRacingTryTimeInputMessage() {
+        System.out.println("시도할 횟수는 몇회인가요?");
+    }
+}
