--- src/main/java/racingcar/input/carname/validator/exception/DuplicateCarNamesException.java
@@ -1,7 +1,0 @@
-package racingcar.input.carname.validator.exception;
-
-public class DuplicateCarNamesException extends RuntimeException {
-    public DuplicateCarNamesException(String message) {
-        super(message);
-    }
-}
