--- src/main/java/racingcar/input/carname/validator/exception/InvalidCarNameFormatException.java
@@ -1,7 +1,0 @@
-package racingcar.input.carname.validator.exception;
-
-public class InvalidCarNameFormatException extends RuntimeException {
-    public InvalidCarNameFormatException(String message) {
-        super(message);
-    }
-}
