--- src/main/java/racingcar/input/racingtrytime/printer/RacingTryTimeInputFromUserPrinter.java
@@ -1,7 +1,0 @@
-package racingcar.input.racingtrytime.printer;
-
-public class RacingTryTimeInputFromUserPrinter {
-    public void printRacingTryTimeInputMessage() {
-        System.out.println("시도할 횟수는 몇회인가요?");
-    }
-}
