--- src/main/java/racingcar/input/racingtrytime/validator/exception/NotNaturalNumberRacingTimeException.java
@@ -1,7 +1,0 @@
-package racingcar.input.racingtrytime.validator.exception;
-
-public class NotNaturalNumberRacingTimeException extends RuntimeException {
-    public NotNaturalNumberRacingTimeException(String message) {
-        super(message);
-    }
-}
